JAVA.SE.01: Sixth task
Created Notepad and Entry classes;
Added NotepadExample class, that shows an example of usage of these classes;

Some code to work with the array was taken from java.util.ArrayList.
Because it is the best and tested by time implementation, and it is not necessary to reinvent the wheel again.
It seems so to me.